backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.1
python-dotenv==1.0.1
openai==1.30.1
python-Levenshtein==0.25.1
httpx==0.27.0
aiofiles==23.2.1
pydantic==2.7.1
pydantic-settings==2.2.1
supabase
# Para producción con PostgreSQL: asyncpg psycopg2-binary supabase
# Para auth completa: python-jose[cryptography] passlib[bcrypt]
